[refactor] score functions
